apps/ai/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
openai==1.13.3
google-generativeai==0.4.1
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
httpx==0.27.0
pydantic==2.6.3
python-multipart==0.0.9
requests==2.31.0
